Add size info text below ImageView
This text displays the size of the ImageView, the size of the original
image, the size of the scaled (displayed image), and the transformation
matrix.

The size values cannot be obtained until the Fragment is completely
created (i.e. not in onCreateView or onResume if the Fragment is freshly
created. That's why there is a button "Measure" to trigger querying these
values and displaying them.