Add app bar action for picking an image
The action is displayed as an icon in the app bar. On clicking, the built-in
Android file explorer is opened, and directories containing images are
displayed (files other than images are greyed out). When the use clicks on
an image, the URI of this image is returned in the Intent in onActivityResult.

Next step: save this URI in the SharedPreferences, and add a listener to the
Fragments, so that when the image changes, all Fragments load the new image
into their ImageViews.