Store image URI in SharedPreferences
The image URI is stored in the shared preferences, and all the fragments
load the current image in onResume.

When the app is restarted, the user-selected image that is currently in the
shared preferences is loaded.

The URI returned by the "choose image" activity has a URI permission. This
is handled differently for pre and post API level 19 devices. In any case,
each time before an image is loaded, it is checked whether we have read
permission on this URI. If yes, the image is loaded, if no, the default
image is loaded.