Validation of EditImageViewActivity text field entries
Idea: separate class with static validation methods for different text
field types. References to these methods are saved as lambda expressions
in a Map, together with the corresponding text fields. It's then possible
to loop through this map and call the appropriate valication method on
the text of each text field.

Lambda expressions are a Java 8 features and can be used with Android
with the new Jack compiler. However, if the Jack compiler is used, then
the Data Binding Library cannot be used (bug that will be fixed in
Android Studio 2.3).

Another option is the Retrolambda library, which allows to use lambda
expressions with Java 7 and 6. This is the approach that is currently
implemented in this commit.

Problem: does not run on API level < 24