Fix API level < 24 problem; UI for validation
Regarding lambda expressions, what does not work on API level < 24 are
the standard Java functional interfaces in java.util.function (both,
with the Jack compiler and Java 8, and with Retrolambda). This problem
can be avoided by defining an own functional interface instead of
using a standard one from java.util.function. In this way, lambda
expressions work flawlessly on API levels < 24 (with Retrolambda).

Started implementing UI for text field validation in EditImageView
Activity.

Current approach: edit field while it is edited or on loss of focus.