UI for validation; validation for "background" field
Finished validation UI so that the text that is currently typed is red
if it is not a valid input for the corresponding field, and turns black
as soon as it is  a valid input.

Not sure whether to replace the dropdown icon in the EditTexts with a
custom AutoCompleteTextView (so that when typing "1", the auto complete
proposes "1dp", "1sp", "1px", "1in", "1mm").