Fix fragment recreation upon returning from EditImageViewActivity
The fragment used in the adapter of the ViewPager is now a special
launcher fragment. This launcher fragment does nothing else than in
onStart launching the real fragment. This includes the creation of
the real fragment from scratch, i.e. starting from onAttach.

In this way, onCreateView is called for the real fragments of the
active two or three pages after returning from EditImageViewActivity
and all the new settings for the ImageView are correctly applied.

Note that the lifetime of a fragment w.r.t. starting another Activity
and returning from it is onPause -> onStop -> onStart -> onResume,
which does not include onCreateView.