Solve view measurement problem in fragments
The problem was that the size of the ImageView, and of its contained
Drawable, could not be measured in the onCreateView, onActivityCreated,
onStart, or onResume callbacks, because at these times the layout,
including the ImageView, has not yet been drawn.

The solution is to add an OnGlobalLayoutListener to the ViewTreeObserver
of the ImageView. The callback method of this listener is then called
multiple times (whenever the layout state of the ImageView changes),
including when the sizes of the ImageView and its contained Drawable are
finally available.